# doc/api.yml
swagger: '2.0'
...
info:
  version: "0.0.0"
  title: AlantuJS REST API
basePath: /api/v1
consumes:
  - application/json
produces:
  - application/json

################################################################################
#                                   Paths                                      #
################################################################################
paths:
...
  /event:
    get:
      description: |
        List all 'Events'
      responses:
        200:
          description:  List all events
          schema:
            title: events
            type: array
            items:
              $ref: '#/definitions/Event'
    post:
      description: |
        Create a 'Event' object.
      parameters:
        - name: principal_id
          in: query
          type: integer
          description: Principal which owns event
          required: true
        - name: user_id
          in: query
          type: integer
          description: User who owns event
          required: true
        - name: event_type_id
          in: query
          type: integer
          description: Event type for event
          required: true
        - name: event
          in: body
          description: The event JSON you want to post
          schema:
            $ref: '#/definitions/Event'
          required: true
      responses:
        200:
          description: Create an event.
  /event/{eventId}:
    get:
      description: |
        Get `Event` object.
      parameters:
        - name: eventId
          in: path
          type: integer
          description: Event identifier
          required: true
      responses:
        200:
          description: |
            Get an events
          schema:
            title: event
            type: object
            items:
              $ref: '#/definitions/Event'
    put:
      description: |
        Update `Event` object.
      parameters:
        - name: eventId
          in: path
          type: integer
          description: Event identifier
          required: true
      responses:
        200:
          description:  |
            Update an existing event.
          schema:
            title: event
            type: object
            items:
              $ref: '#/definitions/Event'
    delete:
      description: |
        Delete an 'Event'.
      parameters:
        - name: eventId
          in: path
          type: integer
          description: Event identifier
          required: true
      responses:
        200:
          description: Deletes an event.

  # Admission REST endpoints
...
  /event_location:
    get:
      description: |
        List all 'Event Location'
      responses:
        200:
          description:  List all Event Locations
          schema:
            title: event_locations
            type: array
            items:
              $ref: '#/definitions/EventLocation'
    post:
      description: |
        Create a 'Event Location' object.
      parameters:
        - name: principal_id
          in: query
          description: Principal associated with event location
          required: true
          type: integer
        - name: event_location
          in: body
          description: The event location JSON you want to post
          schema:
            $ref: '#/definitions/EventLocation'
          required: true
      responses:
        200:
          description: Create a event location.
  /event_location/{eventLocationId}:
    get:
      description: |
        Get `Event Location` object.
      parameters:
        - name: eventLocationId
          in: path
          type: integer
          description: Event location identifier
          required: true
      responses:
        200:
          description: |
            Get a event location
          schema:
            title: event_location
            type: object
            items:
              $ref: '#/definitions/EventLocation'
    put:
      description: |
        Update `Event Location` object.
      parameters:
        - name: eventLocationId
          in: path
          type: integer
          description: Event location identifier
          required: true
        - name: event_location
          in: body
          description: The event location JSON you want to post
          schema:
            $ref: '#/definitions/EventLocation'
          required: true
      responses:
        200:
          description:  |
            Update an existing event location.
          schema:
            title: event_location
            type: object
            items:
              $ref: '#/definitions/EventLocation'
    delete:
      description: |
        Delete a 'Event Location'.
      parameters:
        - name: eventLocationId
          in: path
          type: integer
          description: Event location identifier
          required: true
      responses:
        200:
          description: Deletes a event location.

################################################################################
#                                 Definitions                                #
################################################################################
definitions:
  Event:
    type: object
    properties:
      id:
        type: integer
        format: int32
      principal_id:
        type: integer
        format: int32
      user_account_id:
        type: integer
        format: int32
      event_type_id:
        type: integer
        format: int32
      title:
        type: string
      notes:
        type: string
      location:
        type: string
      preview_img:
        type: string
      created_at:
        type: integer
        format: int64
      updated_at:
        type: integer
        format: int64
      deleted_at:
        type: integer
        format: int64
      event_background:
        type: string
      login_required:
        type: boolean
      latitude:
        type: number
        format: double
      longitude:
        type: number
        format: double
      date:
        type: integer
        format: int64
...
  EventLocation:
    type: object
    properties:
      id:
        type: integer
        format: int32
      principal_id:
        type: integer
        format: int32
      title:
        type: string
      remark:
        type: string
      street:
        type: string
      zip:
        type: string
      city:
        type: string
      opening_hours:
        type: string